Fix auto-paragraphing of the TinyMCE editor.
Because of the way data is extracted from the editor when submitting
edits or new shortcodes, TinyMCE was not running the wpautop
functionality it normally does.  The net effect was newlines in the
visual editors not behaving as expected.

To fix this, I had to run wpautop manually after retrieving the data.