Fix asynchronous JavaScript plugin conflicts.
Some asynchronous JavaScript plugins load ABD's script before jQuery.
Before, the whole world imploded when this happened (or the plugin
failed... one or the other).  Now, ABD at least sits around for a few
seconds to see if jQuery comes around.  On the off chance it doesn't, it
fails a little more gracefully and informatively.